add Edge Rate Limiting functionality in new `erl` package
This package includes a RateCounter type that can be used to
increment an event counter, and to examine the rate of events per
second within a POP over 1, 10, and 60 second windows.  It can also
estimate the number of events seen over the past minute within a POP
in 10 second buckets.

The PenaltyBox type can be used to track entries that should be
penalized for a certain amount of time.

The RateLimiter type combines a rate counter and a penalty box to
determine whether a given entry should be rate limited based on
whether it exceeds a maximum threshold of events per second over a
given rate window.  Most users can simply use RateLimiter.CheckRate
rather than methods on RateCounter and PenaltyBox.

Rate counters and penalty boxes are combined and synchronized within
a POP.  However, Edge Rate Limiting is not intended to compute counts
or rates with high precision and may under count by up to 10%.

Both rate counters and penalty boxes have a fixed capacity for
entries.  Once a rate counter is full, each new entry evicts the
entry that was least recently incremented.  Once a penalty box is
full, each new entry will evict the entry with the shortest remaining
time-to-live (TTL).